requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.21
pyodbc==4.0.39
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.12.0
python-dateutil==2.8.2
python-multipart==0.0.6